+++ packages/cubejs-schema-compiler/src/scaffolding/ScaffoldingTemplate.ts
@@ -1,34 +1,44 @@
 import inflection from 'inflection';
-import { ScaffoldingSchema } from './ScaffoldingSchema';
-import { UserError } from '../compiler/UserError';
+import { CubeDescriptor, CubeDescriptorMember, MemberType, ScaffoldingSchema, TableName } from './ScaffoldingSchema';
+import { UserError } from '../compiler';
 import { ValueWithComments } from './ValueWithComments';
 
+type SchemaContext = {
+  dataSource?: string;
+};
+
+type CubeMembers = {
+  measures: CubeDescriptorMember[];
+  dimensions: CubeDescriptorMember[];
+};
+
 class MemberReference {
-  constructor(member) {
+  public constructor(public member) {
     this.member = member;
   }
 }
 
 export class ScaffoldingTemplate {
-  constructor(dbSchema, driver) {
-    this.dbSchema = dbSchema;
+  private readonly scaffoldingSchema: ScaffoldingSchema;
+
+  public constructor(private readonly dbSchema, private readonly driver) {
     this.scaffoldingSchema = new ScaffoldingSchema(dbSchema);
-    this.driver = driver;
   }
 
-  escapeName(name) {
+  protected escapeName(name) {
     if (this.eligibleIdentifier(name)) {
       return name;
     }
     return this.driver.quoteIdentifier(name);
   }
 
-  eligibleIdentifier(name) {
+  protected eligibleIdentifier(name: string) {
     return !!name.match(/^[a-z0-9_]+$/);
   }
 
-  generateFilesByTableNames(tableNames, schemaContext = {}) {
+  public generateFilesByTableNames(tableNames: TableName[], schemaContext: SchemaContext = {}) {
     const schemaForTables = this.scaffoldingSchema.generateForTables(tableNames.map(n => this.resolveTableName(n)));
+
     return schemaForTables.map(tableSchema => ({
       // eslint-disable-next-line prefer-template
       fileName: tableSchema.cube + '.js',
@@ -36,8 +46,35 @@ export class ScaffoldingTemplate {
     }));
   }
 
+  public generateFilesByCubeDescriptors(cubeDescriptors: CubeDescriptor[], schemaContext = {}) {
+    const tableNames = cubeDescriptors.map(({ tableName }) => tableName);
+    const generatedSchemaForTables = this.scaffoldingSchema.generateForTables(tableNames.map(n => this.resolveTableName(n)));
+
+    const schemaForTables = cubeDescriptors.map((descriptor) => {
+      const generatedDescriptor = generatedSchemaForTables.find(({ cube }) => descriptor.cube);
+      const cubeMembers = descriptor.members.reduce<CubeMembers>((memo, member) => ({
+        measures: [...memo.measures].concat(member.memberType === MemberType.Measure ? [member] : []),
+        dimensions: [...memo.dimensions].concat(member.memberType === MemberType.Dimension ? [member] : []),
+      }), {
+        measures: [],
+        dimensions: []
+      });
+
+      return {
+        ...descriptor,
+        ...generatedDescriptor,
+        ...cubeMembers
+      };
+    });
+
+    return schemaForTables.map(tableSchema => ({
+      fileName: `${tableSchema.cube}.js`,
+      content: this.renderFile(this.schemaDescriptorForTable(tableSchema, schemaContext))
+    }));
+  }
+
   // eslint-disable-next-line consistent-return
-  resolveTableName(tableName) {
+  protected resolveTableName(tableName: TableName) {
     let tableParts;
     if (Array.isArray(tableName)) {
       tableParts = tableName;
@@ -48,7 +85,7 @@ export class ScaffoldingTemplate {
     if (tableParts.length === 2) {
       this.scaffoldingSchema.resolveTableDefinition(tableName);
       return tableName;
-    } else if (tableParts.length === 1) {
+    } else if (tableParts.length === 1 && typeof tableName === 'string') {
       const schema = Object.keys(this.dbSchema).find(
         tableSchema => this.dbSchema[tableSchema][tableName] ||
           this.dbSchema[tableSchema][inflection.tableize(tableName)]
@@ -62,12 +99,12 @@ export class ScaffoldingTemplate {
       if (this.dbSchema[schema][inflection.tableize(tableName)]) {
         return `${schema}.${inflection.tableize(tableName)}`;
       }
-    } else {
-      throw new UserError('Table names should be in <table> or <schema>.<table> format');
     }
+
+    throw new UserError('Table names should be in <table> or <schema>.<table> format');
   }
 
-  schemaDescriptorForTable(tableSchema, schemaContext = {}) {
+  public schemaDescriptorForTable(tableSchema, schemaContext: SchemaContext = {}) {
     return {
       cube: tableSchema.cube,
       sql: `SELECT * FROM ${tableSchema.schema && tableSchema.schema.length ? `${this.escapeName(tableSchema.schema)}.` : ''}${this.escapeName(tableSchema.table)}`, // TODO escape
@@ -84,7 +121,7 @@ export class ScaffoldingTemplate {
       measures: tableSchema.measures.map(m => ({
         [this.memberName(m)]: {
           sql: this.sqlForMember(m),
-          type: m.types[0],
+          type: m.type ?? m.types[0],
           title: this.memberTitle(m)
         }
       })).reduce((a, b) => ({ ...a, ...b }), {
@@ -96,7 +133,7 @@ export class ScaffoldingTemplate {
       dimensions: tableSchema.dimensions.map(m => ({
         [this.memberName(m)]: {
           sql: this.sqlForMember(m),
-          type: m.types[0],
+          type: m.type ?? m.types[0],
           title: this.memberTitle(m),
           primaryKey: m.isPrimaryKey ? true : undefined
         }
@@ -105,25 +142,25 @@ export class ScaffoldingTemplate {
     };
   }
 
-  sqlForMember(m) {
+  protected sqlForMember(m) {
     // eslint-disable-next-line no-template-curly-in-string
     return `${this.escapeName(m.name) !== m.name || !this.eligibleIdentifier(m.name) ? '${CUBE}.' : ''}${this.escapeName(m.name)}`;
   }
 
-  memberTitle(m) {
+  protected memberTitle(m) {
     return inflection.titleize(inflection.underscore(this.memberName(m))) !== m.title ? m.title : undefined;
   }
 
-  memberName(member) {
+  protected memberName(member) {
     return inflection.camelize(member.title.replace(/[^A-Za-z0-9]+/g, '_').toLowerCase(), true);
   }
 
-  renderFile(fileDescriptor) {
+  protected renderFile(fileDescriptor) {
     const { cube, ...descriptor } = fileDescriptor;
     return `cube(\`${cube}\`, ${this.render(descriptor, 0)});\n`;
   }
 
-  render(descriptor, level, appendComment = '') {
+  protected render(descriptor, level, appendComment = '') {
     // eslint-disable-next-line prefer-template
     const lineSeparator = ',\n' + (level < 2 ? '\n' : '');
     if (Array.isArray(descriptor)) {